// types/contracts/GOFPFacet.ts
export interface GOFPFacet extends BaseContract {
  constructor(
    jsonInterface: any[],
    address?: string,
    options?: ContractOptions
  ): GOFPFacet;
  clone(): GOFPFacet;
  methods: {
    adminTerminusInfo(): NonPayableTransactionObject<{
      0: string;
      1: string;
    }>;
​
    chooseCurrentStagePaths(
      sessionId: number | string | BN,
      tokenIds: (number | string | BN)[],
      paths: (number | string | BN)[]
// ...
  };
// ...
}